+++ Sources/GitHub/Model.swift
@@ -2,7 +2,7 @@ import PointFreePrelude
 import Tagged
 
 public struct AccessToken: Codable {
-  public private(set) var accessToken: String
+  public var accessToken: String
 
   private enum CodingKeys: String, CodingKey {
     case accessToken = "access_token"
@@ -10,9 +10,9 @@ public struct AccessToken: Codable {
 }
 
 public struct OAuthError: Codable {
-  public private(set) var description: String
-  public private(set) var error: Error
-  public private(set) var errorUri: String
+  public var description: String
+  public var error: Error
+  public var errorUri: String
 
   public enum Error: String, Codable {
     /// <https://developer.github.com/apps/managing-oauth-apps/troubleshooting-oauth-app-access-token-request-errors/#bad-verification-code>
@@ -27,20 +27,20 @@ public struct OAuthError: Codable {
 }
 
 public struct User: Codable {
-  public private(set) var id: Id
-  public private(set) var name: String?
+  public var id: Id
+  public var name: String?
 
   public struct Email: Codable {
-    public private(set) var email: EmailAddress
-    public private(set) var primary: Bool
+    public var email: EmailAddress
+    public var primary: Bool
   }
 
   public typealias Id = Tagged<User, Int>
 }
 
 public struct UserEnvelope: Codable {
-  public private(set) var accessToken: AccessToken
-  public private(set) var gitHubUser: User
+  public var accessToken: AccessToken
+  public var gitHubUser: User
 
   public init(accessToken: AccessToken, gitHubUser: User) {
     self.accessToken = accessToken
