+++ Sources/PointFree/Environment.swift
@@ -1,9 +1,10 @@
 import Css
 import Foundation
 import GitHub
+import Html
 import Logger
 import Prelude
-import Html
+import Stripe
 
 public var Current = Environment()
 
@@ -20,6 +21,6 @@ public struct Environment {
   public var logger = Logger()
   public var mailgun = Mailgun.live
   public var renderHtml: ([Node]) -> String = Html.render
-  public var stripe = Stripe.live
+  public var stripe = Stripe.Client(secretKey: "")
   public var uuid: () -> UUID = UUID.init
 }
