+++ Sources/PointFreeTestSupport/PointFreeTestSupport.swift
@@ -14,6 +14,7 @@ import Optics
 import PointFreePrelude
 import Prelude
 import SnapshotTesting
+@testable import Stripe
 #if os(macOS)
 import WebKit
 #endif
@@ -195,228 +196,6 @@ extension Pricing {
   public static let teamYearly = mock
     |> \.billing .~ .yearly
     |> \.quantity .~ 4
-}
-
-extension Stripe {
-  public static let mock = Stripe(
-    cancelSubscription: const(pure(.canceling)),
-    createCustomer: { _, _, _ in pure(.mock) },
-    createSubscription: { _, _, _, _ in pure(.mock) },
-    fetchCoupon: const(pure(.mock)),
-    fetchCustomer: const(pure(.mock)),
-    fetchInvoice: const(pure(.mock(charge: .right(.mock)))),
-    fetchInvoices: const(pure(.mock([.mock(charge: .right(.mock))]))),
-    fetchPlans: { pure(.mock([.mock])) },
-    fetchPlan: const(pure(.mock)),
-    fetchSubscription: const(pure(.mock)),
-    fetchUpcomingInvoice: const(pure(.upcoming)),
-    invoiceCustomer: const(pure(.mock(charge: .right(.mock)))),
-    updateCustomer: { _, _ in pure(.mock) },
-    updateCustomerExtraInvoiceInfo: { _, _ in pure(.mock) },
-    updateSubscription: { _, _, _, _ in pure(.mock) },
-    js: ""
-  )
-}
-
-extension Stripe.Card {
-  public static let mock = Stripe.Card(
-    brand: .visa,
-    customer: "cus_test",
-    expMonth: 1,
-    expYear: 2020,
-    id: "card_test",
-    last4: "4242"
-  )
-}
-
-extension Stripe.Charge {
-  public static let mock = Stripe.Charge(
-    amount: 17_00,
-    id: "ch_test",
-    source: .mock
-  )
-}
-
-extension Stripe.Customer {
-  public static let mock = Stripe.Customer(
-    businessVatId: nil,
-    defaultSource: "card_test",
-    id: "cus_test",
-    metadata: [:],
-    sources: .mock([.mock])
-  )
-}
-
-extension Stripe.Error {
-  public static let mock = Stripe.Error(
-    message: "Your card has insufficient funds."
-  )
-}
-
-extension Stripe.ErrorEnvelope {
-  public static let mock = Stripe.ErrorEnvelope(
-    error: .mock
-  )
-}
-
-extension Stripe.Event where T == Either<Stripe.Invoice, Stripe.Subscription> {
-  public static var invoice: Stripe.Event<Either<Stripe.Invoice, Stripe.Subscription>> {
-    return .init(
-      data: .init(object: .left(.mock(charge: .left("ch_test")))),
-      id: "evt_test",
-      type: .invoicePaymentFailed
-    )
-  }
-}
-
-extension Stripe.Invoice {
-  public static func mock(charge: Either<Stripe.Charge.Id, Stripe.Charge>?) -> Stripe.Invoice {
-    return Stripe.Invoice(
-      amountDue: 0_00,
-      amountPaid: 17_00,
-      charge: charge,
-      closed: true,
-      customer: "cus_test",
-      date: .mock,
-      discount: nil,
-      id: "in_test",
-      lines: .mock([.mock]),
-      number: "0000000-0000",
-      periodStart: .mock,
-      periodEnd: Date.mock.addingTimeInterval(60 * 60 * 24 * 30),
-      subscription: "sub_test",
-      subtotal: 17_00,
-      total: 17_00
-    )
-  }
-
-  public static let upcoming = mock(charge: .right(.mock))
-    |> \.amountDue .~ 17_00
-    |> \.amountPaid .~ 0
-    |> \.id .~ nil
-}
-
-extension Stripe.LineItem {
-  public static let mock = Stripe.LineItem(
-    amount: 17_00,
-    description: nil,
-    id: "ii_test",
-    plan: .mock,
-    quantity: 1,
-    subscription: "sub_test"
-  )
-}
-
-extension Stripe.ListEnvelope {
-  public static func mock(_ xs: [A]) -> Stripe.ListEnvelope<A> {
-    return .init(
-      data: xs,
-      hasMore: false
-    )
-  }
-}
-
-extension Stripe.Plan {
-  public static let mock = Stripe.Plan(
-    amount: 17_00,
-    created: .mock,
-    currency: .usd,
-    id: .individualMonthly,
-    interval: .month,
-    metadata: [:],
-    name: "Individual Monthly",
-    statementDescriptor: nil
-  )
-
-  public static let individualMonthly = mock
-
-  public static let individualYearly = mock
-    |> \.amount .~ 170_00
-    |> \.id .~ .individualYearly
-    |> \.interval .~ .year
-    |> \.name .~ "Individual Yearly"
-
-  public static let teamMonthly = individualMonthly
-    |> \.amount .~ 16_00
-    |> \.id .~ .teamMonthly
-    |> \.name .~ "Team Monthly"
-
-  public static let teamYearly = individualYearly
-    |> \.amount .~ 160_00
-    |> \.id .~ .teamYearly
-    |> \.name .~ "Team Yearly"
-}
-
-extension Stripe.Subscription {
-  public static let mock = Stripe.Subscription(
-    canceledAt: nil,
-    cancelAtPeriodEnd: false,
-    created: .mock,
-    currentPeriodStart: .mock,
-    currentPeriodEnd: Date(timeInterval: 60 * 60 * 24 * 30, since: .mock),
-    customer: .right(.mock),
-    discount: nil,
-    endedAt: nil,
-    id: "sub_test",
-    items: .mock([.mock]),
-    plan: .mock,
-    quantity: 1,
-    start: .mock,
-    status: .active
-  )
-
-  public static let individualMonthly = mock
-    |> \.plan .~ .individualMonthly
-    |> \.quantity .~ 1
-
-  public static let individualYearly = mock
-    |> \.plan .~ .individualYearly
-    |> \.quantity .~ 1
-
-  public static let teamMonthly = mock
-    |> \.plan .~ .teamMonthly
-    |> \.quantity .~ 4
-
-  public static let teamYearly = mock
-    |> \.plan .~ .teamYearly
-    |> \.quantity .~ 4
-
-  public static let canceling = mock
-    |> \.cancelAtPeriodEnd .~ true
-
-  public static let canceled = canceling
-    |> \.canceledAt .~ Date(timeInterval: -60 * 60 * 24 * 30, since: .mock)
-    |> \.currentPeriodEnd .~ Date(timeInterval: -60 * 60 * 24 * 30, since: .mock)
-    |> \.currentPeriodStart .~ Date(timeInterval: -60 * 60 * 24 * 60, since: .mock)
-    |> \.status .~ .canceled
-
-  public static let discounted = mock
-    |> \.plan .~ .individualYearly
-    |> \.quantity .~ 1
-    |> \.discount .~ .mock
-}
-
-extension Stripe.Discount {
-  public static let mock = Stripe.Discount(coupon: .mock)
-}
-
-extension Stripe.Coupon {
-  public static let mock = Stripe.Coupon(
-    duration: .forever,
-    id: "coupon-deadbeef",
-    name: "Student Discount",
-    rate: .percentOff(50),
-    valid: true
-  )
-}
-
-extension Stripe.Subscription.Item {
-  public static let mock = Stripe.Subscription.Item(
-    created: .mock,
-    id: "si_test",
-    plan: .mock,
-    quantity: 1
-  )
 }
 
 extension SubscribeData {
